Modify files and clean the codes
