+++ src/App.jsx
@@ -4,11 +4,11 @@ import Error from "./ui/Error";
 import Menu from "./features/menu/Menu";
 import { loader as menuLoader } from "./features/menu/MenuLoader";
 import Cart from "./features/cart/Cart";
-import CreateOrder, { action as createOrderAction } from "./features/order/CreateOrder";
-// import { action as createOrderAction } from "./features/order/Action";
-import Order, { loader as OrderLoader } from "./features/order/Order";
+import CreateOrder from "./features/order/CreateOrder";
+import { action as createOrderAction } from "./features/order/Action";
+import Order from "./features/order/Order";
 import AppLayout from "./ui/AppLayout";
-// import OrderLoader from "./features/order/OrderLoader";
+import OrderLoader from "./features/order/OrderLoader";
 
 const router = createBrowserRouter([
   {
