+++ src/features/order/Action.jsx
@@ -1,29 +1,29 @@
-// import { createOrder } from "../../services/apiRestaurant";
-// import { redirect } from "react-router-dom";
+import { createOrder } from "../../services/apiRestaurant";
+import { redirect } from "react-router-dom";
 
-// export async function action({ request }) {
-//     const formData = await request.formData();
-//     const data = Object.fromEntries(formData);
+export async function action({ request }) {
+    const formData = await request.formData();
+    const data = Object.fromEntries(formData);
 
-//     // https://uibakery.io/regex-library/phone-number
-//     const isValidPhone = (str) =>
-//   /^\+?\d{1,4}?[-.\s]?\(?\d{1,3}?\)?[-.\s]?\d{1,4}[-.\s]?\d{1,4}[-.\s]?\d{1,9}$/.test(
-//     str
-//   );
+    // https://uibakery.io/regex-library/phone-number
+    const isValidPhone = (str) =>
+  /^\+?\d{1,4}?[-.\s]?\(?\d{1,3}?\)?[-.\s]?\d{1,4}[-.\s]?\d{1,4}[-.\s]?\d{1,9}$/.test(
+    str
+  );
   
-//     const order = {
-//       ...data,
-//       cart: JSON.parse(data.cart),
-//       priority: data.priority === "on",
-//     };
+    const order = {
+      ...data,
+      cart: JSON.parse(data.cart),
+      priority: data.priority === "on",
+    };
 
-//     const errors = {};
-//     if(!isValidPhone(order.phone)) errors.phone = 'Please give us your correct phone number. We might need it to contact you.';
+    const errors = {};
+    if(!isValidPhone(order.phone)) errors.phone = 'Please give us your correct phone number. We might need it to contact you.';
 
-//     if (Object.keys(errors).length > 0) return errors;
+    if (Object.keys(errors).length > 0) return errors;
 
-//     const newOrder = await createOrder(order);
+    const newOrder = await createOrder(order);
   
   
-//     return redirect(`/order/${newOrder.id}`);
-//   }
+    return redirect(`/order/${newOrder.id}`);
+  }
