+++ src/features/order/CreateOrder.jsx
@@ -1,9 +1,5 @@
-// import { useState } from "react";
-
-import { Form, redirect, useActionData, useNavigation } from "react-router-dom";
+import { Form, useActionData, useNavigation } from "react-router-dom";
 import Button from "../../ui/Button";
-import { createOrder } from "../../services/apiRestaurant";
-// import { createOrder } from "../../services/apiRestaurant";
 
 // https://uibakery.io/regex-library/phone-number
 const isValidPhone = (str) =>
@@ -101,31 +97,6 @@ function CreateOrder() {
       </Form>
     </div>
   );
-}
-
-export async function action({ request }) {
-  const formData = await request.formData();
-  const data = Object.fromEntries(formData);
-
-  const order = {
-    ...data,
-    cart: JSON.parse(data.cart),
-    priority: data.priority === 'true',
-  };
-
-  console.log(order);
-
-  const errors = {};
-  if (!isValidPhone(order.phone))
-    errors.phone =
-      'Please give us your correct phone number. We might need it to contact you.';
-
-  if (Object.keys(errors).length > 0) return errors;
-
-  // If everything is okay, create new order and redirect
-  const newOrder = await createOrder(order);
-
-  return redirect(`/order/${newOrder.id}`);
 }
 
 export default CreateOrder;
